Enforce noise of type float32

+++ omnisafe/algorithms/sagui/ddpg_lag_randact.py
@@ -125,7 +125,7 @@ class OffPolicyAdapter(OnlineAdapter):
                 act = agent.step(self._current_obs, deterministic=False)
 
                 # Add noise to the action
-                noise = np.random.uniform(-0.12, 0.12, size=act.shape)
+                noise = np.random.uniform(-0.12, 0.12, size=act.shape, dtype=np.float32)
                 act += noise
 
             next_obs, reward, cost, terminated, truncated, info = self.step(act)
